Updating Spring Security, UserRepository

+++ src/main/java/com/nnk/springboot/controllers/BidListController.java
@@ -3,7 +3,6 @@ package com.nnk.springboot.controllers;
 import com.nnk.springboot.domain.BidList;
 import com.nnk.springboot.services.BidListService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -32,12 +31,6 @@ public class BidListController {
         return mav;
     }
 
-//    @PostMapping("/bidList/validate")
-//    public String validate(@Valid BidList bid, BindingResult result, Model model) {
-//        // TODO: check data valid and save to db, after saving return bid list
-//        return "bidList/add";
-//    }
-
     @PostMapping("/bidList/validate")
     public ModelAndView validate(@Valid BidList bid, BindingResult result, Model model) {
         ModelAndView mav = new ModelAndView();
@@ -51,10 +44,9 @@ public class BidListController {
         return mav;
     }
 
-    //TODO: add validation that user != null? Show error message that user doesn't exist?
+    //TODO: Show error message if bid doesn't exist?
     @GetMapping("/bidList/update/{id}")
     public ModelAndView showUpdateForm(@PathVariable("id") Integer id, Model model) {
-        // TODO: get Bid by Id add to model then show to the form
         ModelAndView mav = new ModelAndView();
         BidList bidList = bidListService.findById(id);
         if (bidList != null) {
@@ -65,6 +57,7 @@ public class BidListController {
         return mav;
     }
 
+    //TODO: Add error message if bid doesn't exist?
     @PostMapping("/bidList/update/{id}")
     public ModelAndView updateBid(@PathVariable("id") Integer id, @Valid BidList bidList,
                          BindingResult result, Model model) {
@@ -74,16 +67,22 @@ public class BidListController {
             return mav;
         }
         bidListService.updateBidList(bidList);
+        model.addAttribute("bidList", bidListService.findAllBids());
         mav.setViewName("redirect:/bidList/list");
         return mav;
     }
 
+    //TODO: Add error message if bid doesn't exist?
     @GetMapping("/bidList/delete/{id}")
     public ModelAndView deleteBid(@PathVariable("id") Integer id, Model model) {
-        // TODO: Find Bid by Id and delete the bid, return to Bid list
         ModelAndView mav = new ModelAndView();
-        bidListService.deleteBidList(id);
-        mav.setViewName("redirect:/bidList/list");
+        BidList bidList = bidListService.findById(id);
+        if (bidList != null) {
+            bidListService.deleteBidList(id);
+            model.addAttribute("bidList", bidListService.findAllBids());
+            mav.setViewName("redirect:/bidList/list");
+            return mav;
+        }
         return mav;
     }
 }
